refactor: restructure constants, storage, types
* move into separate dirs for global, plugins, and snippets respectively
