+++ src/utils/global/types.ts
@@ -1,0 +1,35 @@
+export interface Settings {
+  global: boolean;
+  plugins: boolean;
+  themes: boolean;
+  snippets: boolean;
+
+  themeFlavour: string;
+  themeAccent: string;
+  themeLogo: LogoId;
+
+  userSnippets: Record<string, UserSnippet>;
+
+  // Schoolbox URLs
+  urls: string[];
+
+  // whether the settings require a refresh
+  needsRefresh: boolean;
+}
+
+export type LogoId = "default" | "catppuccin" | "schoolbox" | "schooltape" | "schooltape-rainbow" | "schooltape-legacy";
+
+export interface LogoInfo {
+  name: string;
+  url: string;
+  disable?: boolean; // whether the icon should be injected or not
+  adaptive?: boolean; // whether the icon should follow the accent colour
+}
+
+export interface UserSnippet {
+  author: string;
+  name: string;
+  description: string;
+  url: string;
+  toggle: boolean;
+}
