+++ src/utils/plugins/constants.ts
@@ -1,4 +1,4 @@
-import { LogoId, LogoInfo, PluginId, PluginInfo, SnippetId, SnippetInfo } from "./types";
+import { PluginId, PluginInfo } from "./types";
 
 export const PLUGIN_INFO: Record<PluginId, PluginInfo> = {
   subheader: {
@@ -40,47 +40,5 @@ export const PLUGIN_INFO: Record<PluginId, PluginInfo> = {
     name: "Legacy Timetable",
     description: "Moves the timetable to its own row",
     order: 7,
-  },
-};
-
-export const SNIPPET_INFO: Record<SnippetId, SnippetInfo> = {
-  hidePfp: {
-    name: "Hide PFP",
-    description: "Hide your profile picture across Schoolbox.",
-    order: 0,
-  },
-  censor: {
-    name: "Censor",
-    description: "Censors all text and images. This is intended for development purposes.",
-    order: 1,
-  },
-};
-
-export const LOGO_INFO: Record<LogoId, LogoInfo> = {
-  default: {
-    name: "Default",
-    url: "default",
-    disable: true,
-  },
-  catppuccin: {
-    name: "Catppuccin",
-    url: "https://raw.githubusercontent.com/catppuccin/catppuccin/main/assets/logos/exports/1544x1544_circle.png",
-  },
-  schoolbox: {
-    name: "Schoolbox",
-    url: "schoolbox.svg",
-    adaptive: true,
-  },
-  schooltape: {
-    name: "Schooltape",
-    url: "https://schooltape.github.io/schooltape.svg",
-  },
-  "schooltape-rainbow": {
-    name: "ST Rainbow",
-    url: "https://schooltape.github.io/schooltape-ctp.svg",
-  },
-  "schooltape-legacy": {
-    name: "ST Legacy",
-    url: "https://schooltape.github.io/schooltape-legacy.svg",
   },
 };
